refactor: move /v1/dashboard-data to modules
refs #1272
