+++ apps/api/src/block/routes/index.ts
@@ -1,0 +1,1 @@
+export * from "./blocks/blocks.router";
