+++ apps/api/src/dashboard/controllers/network-capacity/network-capacity.controller.ts
@@ -1,0 +1,13 @@
+import { singleton } from "tsyringe";
+
+import { NetworkCapacityResponse } from "../../http-schemas/network-capacity/network-capacity.schema";
+import { StatsService } from "../../services/stats/stats.service";
+
+@singleton()
+export class NetworkCapacityController {
+  constructor(private readonly statsService: StatsService) {}
+
+  async getNetworkCapacity(): Promise<NetworkCapacityResponse> {
+    return await this.statsService.getNetworkCapacity();
+  }
+}
