+++ apps/api/src/dashboard/index.ts
@@ -1,0 +1,1 @@
+export * from "./routes";
