+++ apps/api/src/dashboard/routes/index.ts
@@ -1,0 +1,3 @@
+export * from "@src/dashboard/routes/dashboard-data/dashboard-data.router";
+export * from "@src/dashboard/routes/network-capacity/network-capacity.router";
+export * from "@src/dashboard/routes/graph-data/graph-data.router";
