+++ apps/api/src/services/db/statsService.ts
@@ -1,85 +1,8 @@
-import { AkashBlock as Block } from "@akashnetwork/database/dbSchemas/akash";
-import { Day } from "@akashnetwork/database/dbSchemas/base";
-import { subHours } from "date-fns";
-import { Op, QueryTypes } from "sequelize";
+import { QueryTypes } from "sequelize";
 
 import { chainDb } from "@src/db/dbConnection";
-import { getGpuUtilization } from "./gpuBreakdownService";
 
-type GraphData = {
-  currentValue: number;
-  compareValue: number;
-  snapshots: { date: Date; value: number }[];
-};
-
-const numberOrZero = (x: number | undefined | null) => (typeof x === "number" ? x : 0);
-export const getDashboardData = async () => {
-  const latestBlockStats = await Block.findOne({
-    where: {
-      isProcessed: true,
-      totalUUsdSpent: { [Op.not]: null }
-    },
-    order: [["height", "DESC"]]
-  });
-  if (!latestBlockStats) {
-    throw new Error("No blocks stats found");
-  }
-
-  const compareDate = subHours(latestBlockStats.datetime, 24);
-  const compareBlockStats = await Block.findOne({
-    order: [["datetime", "ASC"]],
-    where: {
-      datetime: { [Op.gte]: compareDate }
-    }
-  });
-
-  const secondCompareDate = subHours(latestBlockStats.datetime, 48);
-  const secondCompareBlockStats = await Block.findOne({
-    order: [["datetime", "ASC"]],
-    where: {
-      datetime: { [Op.gte]: secondCompareDate }
-    }
-  });
-
-  return {
-    now: {
-      date: latestBlockStats.datetime,
-      height: latestBlockStats.height,
-      activeLeaseCount: latestBlockStats.activeLeaseCount,
-      totalLeaseCount: latestBlockStats.totalLeaseCount,
-      dailyLeaseCount: numberOrZero(latestBlockStats.totalLeaseCount) - numberOrZero(compareBlockStats?.totalLeaseCount),
-      totalUAktSpent: latestBlockStats.totalUAktSpent,
-      dailyUAktSpent: numberOrZero(latestBlockStats.totalUAktSpent) - numberOrZero(compareBlockStats?.totalUAktSpent),
-      totalUUsdcSpent: latestBlockStats.totalUUsdcSpent,
-      dailyUUsdcSpent: numberOrZero(latestBlockStats.totalUUsdcSpent) - numberOrZero(compareBlockStats?.totalUUsdcSpent),
-      totalUUsdSpent: latestBlockStats.totalUUsdSpent,
-      dailyUUsdSpent: numberOrZero(latestBlockStats.totalUUsdSpent) - numberOrZero(compareBlockStats?.totalUUsdSpent),
-      activeCPU: latestBlockStats.activeCPU,
-      activeGPU: latestBlockStats.activeGPU,
-      activeMemory: latestBlockStats.activeMemory,
-      activeStorage: numberOrZero(latestBlockStats.activeEphemeralStorage) + numberOrZero(latestBlockStats.activePersistentStorage)
-    },
-    compare: {
-      date: compareBlockStats?.datetime,
-      height: compareBlockStats?.height,
-      activeLeaseCount: compareBlockStats?.activeLeaseCount,
-      totalLeaseCount: compareBlockStats?.totalLeaseCount,
-      dailyLeaseCount: numberOrZero(compareBlockStats?.totalLeaseCount) - numberOrZero(secondCompareBlockStats?.totalLeaseCount),
-      totalUAktSpent: compareBlockStats?.totalUAktSpent,
-      dailyUAktSpent: numberOrZero(compareBlockStats?.totalUAktSpent) - numberOrZero(secondCompareBlockStats?.totalUAktSpent),
-      totalUUsdcSpent: compareBlockStats?.totalUUsdcSpent,
-      dailyUUsdcSpent: numberOrZero(compareBlockStats?.totalUUsdcSpent) - numberOrZero(secondCompareBlockStats?.totalUUsdcSpent),
-      totalUUsdSpent: compareBlockStats?.totalUUsdSpent,
-      dailyUUsdSpent: numberOrZero(compareBlockStats?.totalUUsdSpent) - numberOrZero(secondCompareBlockStats?.totalUUsdSpent),
-      activeCPU: compareBlockStats?.activeCPU,
-      activeGPU: compareBlockStats?.activeGPU,
-      activeMemory: compareBlockStats?.activeMemory,
-      activeStorage: numberOrZero(compareBlockStats?.activeEphemeralStorage) + numberOrZero(compareBlockStats?.activePersistentStorage)
-    }
-  };
-};
-
-type AuthorizedGraphDataName =
+export type AuthorizedGraphDataName =
   | "dailyUAktSpent"
   | "dailyUUsdcSpent"
   | "dailyUUsdSpent"
@@ -114,88 +37,6 @@ export const AuthorizedGraphDataNames: AuthorizedGraphDataName[] = [
 
 export function isValidGraphDataName(x: string): x is AuthorizedGraphDataName {
   return AuthorizedGraphDataNames.includes(x as AuthorizedGraphDataName);
-}
-
-export async function getGraphData(dataName: AuthorizedGraphDataName): Promise<GraphData> {
-  let attributes: (keyof Block)[] = [];
-  let isRelative = false;
-  let getter: (block: Block) => number;
-
-  switch (dataName) {
-    case "dailyUAktSpent":
-      attributes = ["totalUAktSpent"];
-      getter = (block: Block) => numberOrZero(block.totalUAktSpent);
-      isRelative = true;
-      break;
-    case "dailyUUsdcSpent":
-      attributes = ["totalUUsdcSpent"];
-      getter = (block: Block) => numberOrZero(block.totalUUsdcSpent);
-      isRelative = true;
-      break;
-    case "dailyUUsdSpent":
-      attributes = ["totalUUsdSpent"];
-      getter = (block: Block) => numberOrZero(block.totalUUsdSpent);
-      isRelative = true;
-      break;
-    case "dailyLeaseCount":
-      attributes = ["totalLeaseCount"];
-      getter = (block: Block) => numberOrZero(block.totalLeaseCount);
-      isRelative = true;
-      break;
-    case "activeStorage":
-      attributes = ["activeEphemeralStorage", "activePersistentStorage"];
-      getter = (block: Block) => numberOrZero(block.activeEphemeralStorage) + numberOrZero(block.activePersistentStorage);
-      break;
-    case "gpuUtilization":
-      return await getGpuUtilization();
-    default:
-      attributes = [dataName];
-      getter = (block: Block) => numberOrZero(block[dataName]);
-  }
-
-  const result = await Day.findAll({
-    attributes: ["date"],
-    include: [
-      {
-        model: Block,
-        as: "lastBlock",
-        attributes: attributes,
-        required: true
-      }
-    ],
-    order: [["date", "ASC"]]
-  });
-
-  let stats = result.map(day => ({
-    date: day.date,
-    value: getter(day.lastBlock!)
-  }));
-
-  if (dataName === "activeGPU") {
-    const firstWithValue = stats.findIndex(x => x.value > 0);
-    stats = stats.filter((_, i) => i >= firstWithValue);
-  }
-
-  if (isRelative) {
-    const relativeStats = stats.reduce<{ date: Date; value: number }[]>((arr, dataPoint, index) => {
-      arr[index] = {
-        date: dataPoint.date,
-        value: dataPoint.value - (index > 0 ? stats[index - 1].value : 0)
-      };
-
-      return arr;
-    }, []);
-
-    stats = relativeStats;
-  }
-
-  const dashboardData = await getDashboardData();
-
-  return {
-    currentValue: numberOrZero(dashboardData.now[dataName]),
-    compareValue: numberOrZero(dashboardData.compare[dataName]),
-    snapshots: stats
-  };
 }
 
 export async function getProviderTotalLeaseCountAtHeight(provider: string, height: number) {
