+++ apps/api/src/transaction/routes/index.ts
@@ -1,0 +1,1 @@
+export * from "./transactions/transactions.router";
