+++ apps/api/test/functional/dashboard-data.spec.ts
@@ -112,27 +112,6 @@ describe("Dashboard Data", () => {
         checkDate: twoDaysAgo,
         isOnline: true,
         isLastSuccessOfDay: true,
-        activeCPU: 301,
-        activeGPU: 302,
-        activeMemory: 303,
-        activePersistentStorage: 304,
-        activeEphemeralStorage: 305,
-        pendingCPU: 306,
-        pendingGPU: 307,
-        pendingMemory: 308,
-        pendingPersistentStorage: 309,
-        pendingEphemeralStorage: 310,
-        availableCPU: 311,
-        availableGPU: 312,
-        availableMemory: 313,
-        availablePersistentStorage: 314,
-        availableEphemeralStorage: 315
-      }),
-      ProviderSnapshotSeeder.createInDatabase({
-        owner: providers[1].owner,
-        checkDate: yesterday,
-        isOnline: true,
-        isLastSuccessOfDay: true,
         activeCPU: 401,
         activeGPU: 402,
         activeMemory: 403,
@@ -151,7 +130,7 @@ describe("Dashboard Data", () => {
       }),
       ProviderSnapshotSeeder.createInDatabase({
         owner: providers[1].owner,
-        checkDate: now,
+        checkDate: yesterday,
         isOnline: true,
         isLastSuccessOfDay: true,
         activeCPU: 501,
@@ -169,6 +148,27 @@ describe("Dashboard Data", () => {
         availableMemory: 513,
         availablePersistentStorage: 514,
         availableEphemeralStorage: 515
+      }),
+      ProviderSnapshotSeeder.createInDatabase({
+        owner: providers[1].owner,
+        checkDate: now,
+        isOnline: true,
+        isLastSuccessOfDay: true,
+        activeCPU: 601,
+        activeGPU: 602,
+        activeMemory: 603,
+        activePersistentStorage: 604,
+        activeEphemeralStorage: 605,
+        pendingCPU: 606,
+        pendingGPU: 607,
+        pendingMemory: 608,
+        pendingPersistentStorage: 609,
+        pendingEphemeralStorage: 610,
+        availableCPU: 611,
+        availableGPU: 612,
+        availableMemory: 613,
+        availablePersistentStorage: 614,
+        availableEphemeralStorage: 615
       })
     ]);
     await Promise.all([
